+++ hmapper/src/worker.h
@@ -20,10 +20,10 @@
  */
 typedef struct __attribute__((__packed__)) meth_array_node {
   uint32_t position;    /* Position in the chromosome */
-  uint8_t c_count;      /* Number of total cytosines  */
-  uint8_t nc_count;     /* Number of non-methylated cytosines */
-  uint8_t mc_count;     /* Number of 5-mC cytosines   */
-  uint8_t hmc_count;    /* Number of 5-hmC cytosines  */
+  uint16_t c_count;      /* Number of total cytosines  */
+  uint16_t nc_count;     /* Number of non-methylated cytosines */
+  uint16_t mc_count;     /* Number of 5-mC cytosines   */
+  uint16_t hmc_count;    /* Number of 5-hmC cytosines  */
 } meth_array_node_t;
